feat(pos): moon chain per live EVE - reactor runs selected invTypeReactions formula (x100 batch), moon composition generated by rarity/security/region (Technetium->Guristas); widen reaction column

+++ src/eve-server/DataClasses.h
@@ -32,15 +32,15 @@ public:
     std::map<uint32, EVEPOS::POS_Connections>& GetConnections()     { return connections; }
     std::map<uint32, EVEPOS::POS_Resource>& GetDemands()            { return demands; }
     std::map<uint32, EVEPOS::POS_Resource>& GetSupplies()           { return supplies; }
-    int16 GetReaction() const                           { return reaction; }
-    void SetReaction(int16 react)                       { reaction = react; }
+    int32 GetReaction() const                           { return reaction; }
+    void SetReaction(int32 react)                       { reaction = react; }
     int32 GetItemID() const                             { return itemID; }
     void SetItemID(int32 id)                            { itemID = id; }
 
 private:
     bool active;
     int32 itemID;
-    int16 reaction;     // bp typeID?
+    int32 reaction;     // reactionTypeID (invTypeReactions) selected for this module
     std::map<uint32, EVEPOS::POS_Connections> connections;  // itemID, data
     std::map<uint32, EVEPOS::POS_Resource> demands;         // itemID, resourceData(typeID/quantity)
     std::map<uint32, EVEPOS::POS_Resource> supplies;        // itemID, resourceData(typeID/quantity)
